pyproject.toml
[project]
name = "agentic_ai_langgraph"
version = "0.1.0"
description = "Agentic AI with LangGraph + LangChain"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-tavily>=0.2.11",
    "langchain-google-genai>=2.1.10",
    "langchain-text-splitters>=0.3.11",
    "langchain-tools>=0.1.34",
    "python-dotenv>=1.0.1",
    "langgraph>=0.6.6",
    "grandalf>=0.8",
    "langsmith>=0.4.23",
]
